pyproject.toml
[project]
name = "mcp_snowflake_server"
version = "0.4.0"
description = "A simple Snowflake MCP server"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "mcp>=1.0.0",
    "snowflake-connector-python[pandas]>=3.12.0,<3.14.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "sqlparse>=0.5.3",
    "snowflake-snowpark-python>=1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389"]

[project.scripts]
mcp_snowflake_server = "mcp_snowflake_server.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_snowflake_server"]
